test(InterventionList): merge test into one file
