+++ frontend/src/__tests__/components/PatientPage/InterventionList.test.tsx
@@ -1,12 +1,15 @@
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import InterventionList from '@/components/PatientPage/InterventionList';
-import apiClient from '@/api/client';
 import '@testing-library/jest-dom';
+import { patientUiStore } from '@/stores/patientUiStore';
+import { patientInterventionsStore } from '@/stores/patientInterventionsStore';
 import { patientQuestionnairesStore } from '@/stores/patientQuestionnairesStore';
+
+// ── Global mocks ──────────────────────────────────────────────────────────────
 
 jest.mock('react-i18next', () => ({
   useTranslation: () => ({
-    t: (key: string) => key, // Mock translations
+    t: (key: string) => key,
     i18n: { language: 'en' },
   }),
 }));
@@ -20,327 +23,88 @@ jest.mock('@/utils/translate', () => ({
   ),
 }));
 
+jest.mock('@/api/client', () => require('@/__mocks__/api/client'));
+
+jest.mock('@/stores/authStore', () => ({
+  __esModule: true,
+  default: { id: 'p1' },
+}));
+
+jest.mock('@/stores/patientUiStore', () => {
+  const store = {
+    selectedDate: new Date('2026-02-16T00:00:00Z'),
+    viewMode: 'day',
+    setSelectedDate: jest.fn(),
+    setViewMode: jest.fn((mode) => {
+      store.viewMode = mode;
+    }),
+    goToday: jest.fn(),
+  };
+  return { patientUiStore: store };
+});
+
+jest.mock('@/stores/patientInterventionsStore', () => ({
+  patientInterventionsStore: {
+    items: [
+      {
+        intervention_id: 'int1',
+        intervention_title: 'Walk',
+        translated_title: 'Walk',
+        translated_description: 'desc',
+        dates: ['2026-02-16T08:00:00.000Z'],
+        duration: 10,
+      },
+    ],
+    error: '',
+    errorDetails: '',
+    fetchPlan: jest.fn(),
+    toggleCompleted: jest.fn(),
+    isCompletedOn: jest.fn(),
+  },
+}));
+
+jest.mock('@/stores/patientQuestionnairesStore', () => ({
+  patientQuestionnairesStore: {
+    showFeedbackPopup: false,
+    feedbackInterventionId: '',
+    feedbackQuestions: [],
+    feedbackDateKey: '',
+    showHealthPopup: false,
+    showInitialPopup: false,
+    healthQuestions: [],
+    closeFeedback: jest.fn(),
+    closeHealth: jest.fn(),
+    closeInitial: jest.fn(),
+    checkInitialQuestionnaire: jest.fn(),
+    loadHealthQuestionnaire: jest.fn(),
+    openInterventionFeedback: jest.fn(),
+  },
+}));
+
+// Reduce popup complexity in tests
+jest.mock('@/components/PatientPage/PatientInterventionPopUp', () => () => (
+  <div data-testid="info-popup" />
+));
+jest.mock('@/components/PatientPage/FeedbackPopup', () => () => (
+  <div data-testid="feedback-popup" />
+));
+jest.mock('@/components/PatientPage/PatientQuestionaire', () => () => (
+  <div data-testid="initial-popup" />
+));
+
+// ── Shared date helpers ───────────────────────────────────────────────────────
+
 const today = new Date();
 const pastDate = new Date(today);
 pastDate.setDate(today.getDate() - 1);
 const futureDate = new Date(today);
 futureDate.setDate(today.getDate() + 1);
-jest.mock('@/api/client', () => require('@/__mocks__/api/client'));
+
+// ── InterventionList Component (UI / rendering tests) ────────────────────────
+
 describe('InterventionList Component', () => {
   beforeEach(() => {
-    // Reset MobX store state so popup state doesn't leak between tests.
-    patientQuestionnairesStore.closeFeedback();
-    patientQuestionnairesStore.closeHealth();
-    patientQuestionnairesStore.closeInitial();
-
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/67d588798c0494979e4633e4/')) {
-        return Promise.resolve({
-          data: [
-            [
-              {
-                intervention_id: '67593dffda564bcba2b249d5',
-                intervention_title: 'Yoga for 60 minutes',
-                description: 'Relaxing morning yoga',
-                frequency: '',
-                dates: [
-                  '2025-04-08T06:00:00',
-                  '2025-04-09T06:00:00',
-                  '2025-04-10T06:00:00',
-                  '2025-04-11T06:00:00',
-                  '2025-04-12T06:00:00',
-                  '2025-04-13T06:00:00',
-                  '2025-04-14T06:00:00',
-                  '2025-04-15T06:00:00',
-                ],
-                completion_dates: [
-                  '2025-04-09T11:24:17.164000',
-                  '2025-04-16T14:52:48.363000',
-                  '2025-04-17T07:14:57.845000',
-                  '2025-04-24T11:31:15.584000',
-                  '2025-04-25T16:12:25.872000',
-                ],
-                content_type: 'Video',
-                benefitFor: [],
-                tags: [],
-                duration: 60,
-                feedback: [],
-                preview_img: '',
-                link: 'https://www.youtube.com/embed/1-lO2U7mxIY?si=7y1nmLDSDrD6wzFt',
-              },
-              {
-                intervention_id: '675a85018ea37a32e90afe5e',
-                intervention_title: '30 minutes meditation',
-                description: 'Meditate for 30 minutes with this audio guide.',
-                frequency: '',
-                dates: [
-                  '2025-04-08T06:00:00',
-                  '2025-04-09T06:00:00',
-                  '2025-04-10T06:00:00',
-                  '2025-04-11T06:00:00',
-                  '2025-04-12T06:00:00',
-                  '2025-04-13T06:00:00',
-                ],
-                completion_dates: [
-                  '2025-04-09T11:37:43.093000',
-                  '2025-04-24T11:31:28.923000',
-                  '2025-04-25T16:12:30.285000',
-                ],
-                content_type: 'Exercise',
-                benefitFor: [],
-                tags: [],
-                duration: null,
-                feedback: [],
-                preview_img: '',
-                media_file:
-                  'http://159.100.246.89:8000/media/audio/20241212063856_TheMusicMeditationPodcast-20240413-BuildCourageAndConfidenceWithGabbyBernstein.mp3',
-              },
-              {
-                intervention_id: '67a48456706dbc891ec91d61',
-                intervention_title: 'Gute Laune Work-Out',
-                description:
-                  'Mach mit bei unserem Gute Laune Workout, um mit Sport gegen Winterdepressionen anzuk\u00e4mpfen!',
-                frequency: '',
-                dates: [
-                  '2025-04-09T08:00:00',
-                  '2025-04-11T08:00:00',
-                  '2025-04-13T08:00:00',
-                  '2025-04-14T08:00:00',
-                  '2025-04-16T08:00:00',
-                  '2025-04-18T08:00:00',
-                  '2025-04-20T08:00:00',
-                  '2025-04-21T08:00:00',
-                ],
-                completion_dates: [],
-                content_type: 'Video',
-                benefitFor: ['Mobility,Psychische', 'Gesundheit'],
-                tags: ['K\u00f6rperliche', 'Aktivit\u00e4t'],
-                duration: 15,
-                feedback: [],
-                preview_img: '',
-                link: 'https://www.youtube.com/watch?v=eC6qEAME7yA',
-              },
-              {
-                intervention_id: '675a86ea8ea37a32e90afe5f',
-                intervention_title: 'What is the new food pyramid?',
-                description: 'Read the article from SRF about the new food pyramid.',
-                frequency: '',
-                dates: ['2025-04-10T06:00:00'],
-                completion_dates: [],
-                content_type: 'Article',
-                benefitFor: [],
-                tags: [],
-                duration: null,
-                feedback: [],
-                preview_img: '',
-                link: 'https://www.srf.ch/radio-srf-1/ernaehrungsempfehlungen-facelifting-der-lebensmittelpyramide-fuenf-wichtige-neuerungen',
-              },
-              {
-                intervention_id: '675a88118ea37a32e90afe67',
-                intervention_title: 'Why sleep is important?',
-                description: 'Read an article about the importance of sleep.',
-                frequency: '',
-                dates: ['2025-04-23T07:00:00'],
-                completion_dates: [],
-                content_type: 'Article',
-                benefitFor: [],
-                tags: [],
-                duration: null,
-                feedback: [],
-                preview_img: '',
-                link: 'https://www.nhlbi.nih.gov/health/sleep/why-sleep-important',
-              },
-            ],
-          ],
-        });
-      }
-      if (url.includes('/patients/get-questions/Healthstatus/67d588798c0494979e4633e4/')) {
-        return Promise.resolve({
-          data: {
-            questions: [
-              {
-                questionKey: 'hs_pain_level',
-                answerType: 'dropdown',
-                translations: [
-                  { language: 'en', text: 'How would you rate your current level of pain?' },
-                  {
-                    language: 'de',
-                    text: 'Wie w\u00fcrden Sie Ihr aktuelles Schmerzlevel einsch\u00e4tzen?',
-                  },
-                ],
-                possibleAnswers: [
-                  {
-                    key: 'no_pain',
-                    translations: [
-                      { language: 'en', text: 'No pain' },
-                      { language: 'de', text: 'Keine Schmerzen' },
-                    ],
-                  },
-                  {
-                    key: 'mild_pain',
-                    translations: [
-                      { language: 'en', text: 'Mild pain' },
-                      { language: 'de', text: 'Leichte Schmerzen' },
-                    ],
-                  },
-                  {
-                    key: 'moderate_pain',
-                    translations: [
-                      { language: 'en', text: 'Moderate pain' },
-                      { language: 'de', text: 'M\u00e4\u00dfige Schmerzen' },
-                    ],
-                  },
-                  {
-                    key: 'severe_pain',
-                    translations: [
-                      { language: 'en', text: 'Severe pain' },
-                      { language: 'de', text: 'Starke Schmerzen' },
-                    ],
-                  },
-                ],
-              },
-              {
-                questionKey: 'hs_energy_level',
-                answerType: 'dropdown',
-                translations: [
-                  { language: 'en', text: 'How energetic do you feel today?' },
-                  { language: 'de', text: 'Wie energiegeladen f\u00fchlen Sie sich heute?' },
-                ],
-                possibleAnswers: [
-                  {
-                    key: 'very_energetic',
-                    translations: [
-                      { language: 'en', text: 'Very energetic' },
-                      { language: 'de', text: 'Sehr energiegeladen' },
-                    ],
-                  },
-                  {
-                    key: 'somewhat_energetic',
-                    translations: [
-                      { language: 'en', text: 'Somewhat energetic' },
-                      { language: 'de', text: 'Etwas energiegeladen' },
-                    ],
-                  },
-                  {
-                    key: 'tired',
-                    translations: [
-                      { language: 'en', text: 'Tired' },
-                      { language: 'de', text: 'M\u00fcde' },
-                    ],
-                  },
-                  {
-                    key: 'exhausted',
-                    translations: [
-                      { language: 'en', text: 'Exhausted' },
-                      { language: 'de', text: 'Ersch\u00f6pft' },
-                    ],
-                  },
-                ],
-              },
-              {
-                questionKey: 'hs_sleep_quality',
-                answerType: 'dropdown',
-                translations: [
-                  { language: 'en', text: 'How would you describe your sleep quality last night?' },
-                  {
-                    language: 'de',
-                    text: 'Wie w\u00fcrden Sie Ihre Schlafqualit\u00e4t letzte Nacht beschreiben?',
-                  },
-                ],
-                possibleAnswers: [
-                  {
-                    key: 'very_good',
-                    translations: [
-                      { language: 'en', text: 'Very good' },
-                      { language: 'de', text: 'Sehr gut' },
-                    ],
-                  },
-                  {
-                    key: 'good',
-                    translations: [
-                      { language: 'en', text: 'Good' },
-                      { language: 'de', text: 'Gut' },
-                    ],
-                  },
-                  {
-                    key: 'fair',
-                    translations: [
-                      { language: 'en', text: 'Fair' },
-                      { language: 'de', text: 'Befriedigend' },
-                    ],
-                  },
-                  {
-                    key: 'poor',
-                    translations: [
-                      { language: 'en', text: 'Poor' },
-                      { language: 'de', text: 'Schlecht' },
-                    ],
-                  },
-                ],
-              },
-              {
-                questionKey: 'hs_emotional_state',
-                answerType: 'multi-select',
-                translations: [
-                  { language: 'en', text: 'Which emotions best describe how you feel today?' },
-                  {
-                    language: 'de',
-                    text: 'Welche Gef\u00fchle beschreiben am besten, wie Sie sich heute f\u00fchlen?',
-                  },
-                ],
-                possibleAnswers: [
-                  {
-                    key: 'happy',
-                    translations: [
-                      { language: 'en', text: 'Happy' },
-                      { language: 'de', text: 'Gl\u00fccklich' },
-                    ],
-                  },
-                  {
-                    key: 'anxious',
-                    translations: [
-                      { language: 'en', text: 'Anxious' },
-                      { language: 'de', text: '\u00c4ngstlich' },
-                    ],
-                  },
-                  {
-                    key: 'frustrated',
-                    translations: [
-                      { language: 'en', text: 'Frustrated' },
-                      { language: 'de', text: 'Frustriert' },
-                    ],
-                  },
-                  {
-                    key: 'calm',
-                    translations: [
-                      { language: 'en', text: 'Calm' },
-                      { language: 'de', text: 'Ruhig' },
-                    ],
-                  },
-                ],
-              },
-              {
-                questionKey: 'hs_mobility_limitations',
-                answerType: 'text',
-                translations: [
-                  {
-                    language: 'en',
-                    text: 'Please describe any limitations you experienced with mobility today.',
-                  },
-                  {
-                    language: 'de',
-                    text: 'Bitte beschreiben Sie eventuelle Einschr\u00e4nkungen Ihrer Mobilit\u00e4t heute.',
-                  },
-                ],
-                possibleAnswers: [],
-              },
-            ],
-          },
-        });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
+    // Set patientId via localStorage so the component can find it
     Object.defineProperty(window, 'localStorage', {
       value: {
         getItem: jest.fn((key) => {
@@ -348,27 +112,15 @@ describe('InterventionList Component', () => {
           return null;
         }),
       },
+      writable: true,
     });
 
     jest.clearAllMocks();
-    // Reset MobX store state that persists between tests
-    patientQuestionnairesStore.closeFeedback();
-    patientQuestionnairesStore.closeHealth();
-    patientQuestionnairesStore.closeInitial();
-  });
-
-  it('renders without crashing and calls API', async () => {
-    (apiClient.get as jest.Mock).mockResolvedValueOnce({ data: [] }); // Mock empty interventions
-    (apiClient.get as jest.Mock).mockResolvedValueOnce({ data: { questions: [] } }); // Mock questions
-
-    render(<InterventionList />);
-
-    await waitFor(() => {
-      expect(apiClient.get).toHaveBeenCalledWith(
-        '/patients/rehabilitation-plan/patient/67d588798c0494979e4633e4/',
-        { params: { lang: 'en' } }
-      );
-    });
+    // Restore isCompletedOn default after clearAllMocks
+    (patientInterventionsStore.isCompletedOn as jest.Mock).mockReturnValue(false);
+    (patientQuestionnairesStore.closeFeedback as jest.Mock).mockImplementation(() => {});
+    (patientQuestionnairesStore.closeHealth as jest.Mock).mockImplementation(() => {});
+    (patientQuestionnairesStore.closeInitial as jest.Mock).mockImplementation(() => {});
   });
 
   it('shows Day and Week toggle buttons', () => {
@@ -377,29 +129,6 @@ describe('InterventionList Component', () => {
     expect(screen.getByLabelText(/Week view/i)).toBeInTheDocument();
   });
 
-  it('calls fetchInterventions on mount', async () => {
-    const mockData = {
-      data: [
-        {
-          intervention_id: '1',
-          intervention_title: 'Test',
-          dates: ['2025-04-28'],
-          description: 'desc',
-          duration: 30,
-          completion_dates: [],
-        },
-      ],
-    };
-    (apiClient.get as jest.Mock).mockResolvedValueOnce(mockData);
-
-    render(<InterventionList />);
-    await waitFor(() =>
-      expect(apiClient.get).toHaveBeenCalledWith(
-        expect.stringContaining('/patients/rehabilitation-plan/patient/'),
-        expect.objectContaining({ params: expect.objectContaining({ lang: expect.any(String) }) })
-      )
-    );
-  });
   it('navigates to next and previous day/week', () => {
     render(<InterventionList />);
     const nextButton = screen.getByText('Next');
@@ -411,462 +140,113 @@ describe('InterventionList Component', () => {
     expect(nextButton).toBeInTheDocument();
     expect(prevButton).toBeInTheDocument();
   });
-  it('displays correct badge status for interventions', async () => {
-    const mockInterventions = [
-      {
-        intervention_id: '1',
-        intervention_title: 'Physio',
-        dates: [new Date().toISOString().split('T')[0]], // Today
-        description: 'desc',
-        duration: 20,
-        completion_dates: [],
-      },
-    ];
 
-    (apiClient.get as jest.Mock).mockResolvedValueOnce({ data: mockInterventions });
+  it('displays the intervention item for the selected date', () => {
     render(<InterventionList />);
-
-    await waitFor(() => {
-      expect(screen.getByText(/Ididit/i)).toBeInTheDocument();
-    });
-  });
-  it('loads health questionnaire if patient ID matches', async () => {
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/67d588798c0494979e4633e4/')) {
-        return Promise.resolve({ data: [] });
-      }
-      if (url.includes('/patients/get-questions/Healthstatus/67d588798c0494979e4633e4/')) {
-        return Promise.resolve({
-          data: {
-            questions: [
-              {
-                questionKey: 'intervention_done',
-                answerType: 'dropdown',
-                translations: [
-                  { language: 'en', text: 'Did you do the excercise fully?' },
-                  { language: 'de', text: 'Haben Sie die Übung vollständig durchgeführt?' },
-                ],
-                possibleAnswers: [
-                  {
-                    key: 'option_1',
-                    translations: [
-                      { language: 'en', text: 'Yes' },
-                      { language: 'de', text: 'Ja' },
-                    ],
-                  },
-                  {
-                    key: 'option_2',
-                    translations: [
-                      { language: 'en', text: 'No' },
-                      { language: 'de', text: 'Nein' },
-                    ],
-                  },
-                ],
-              },
-            ],
-          },
-        });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
-    render(<InterventionList />);
-
-    // The health questionnaire is loaded in the background but doesn't show automatically
-    // Just verify the API was called
-    await waitFor(() => {
-      expect(apiClient.get).toHaveBeenCalledWith(
-        'patients/get-questions/Healthstatus/67d588798c0494979e4633e4/',
-        expect.any(Object)
-      );
-    });
+    // patientInterventionsStore.items has 'Walk' for 2026-02-16 and
+    // patientUiStore.selectedDate is 2026-02-16 (from store mock)
+    expect(screen.getByText(/Walk/i)).toBeInTheDocument();
   });
 
-  it('shows health questionnaire popup on mount if questions exist', async () => {
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({ data: [] });
-      }
-      if (url.includes('/patients/get-questions/Healthstatus/')) {
-        return Promise.resolve({
-          data: {
-            questions: [
-              {
-                questionKey: 'q1',
-                translations: [{ language: 'en', text: 'How do you feel?' }],
-                possibleAnswers: [],
-                answerType: 'text',
-              },
-            ],
-          },
-        });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
+  it('shows "Ididit" button for a past uncompleted intervention', () => {
     render(<InterventionList />);
-
-    // The questionnaire loads but doesn't show automatically - it shows after intervention completion
-    // Just verify the API was called correctly
-    await waitFor(() => {
-      expect(apiClient.get).toHaveBeenCalledWith(
-        'patients/get-questions/Healthstatus/67d588798c0494979e4633e4/',
-        expect.objectContaining({
-          headers: expect.objectContaining({ 'Accept-Language': 'en' }),
-        })
-      );
-    });
+    // Walk item is for 2026-02-16 (past), isCompletedOn = false → Ididit button
+    expect(screen.getByRole('button', { name: /Ididit/i })).toBeInTheDocument();
   });
 
-  it('calls mark as done when clicking "I did it" button', async () => {
-    (apiClient.get as jest.Mock).mockResolvedValueOnce({
-      data: [
-        {
-          intervention_id: '123',
-          intervention_title: 'Exercise 1',
-          description: 'Do something helpful',
-          duration: 30,
-          dates: [new Date().toISOString().split('T')[0]],
-          completion_dates: [],
-        },
-      ],
-    });
-
-    (apiClient.get as jest.Mock).mockResolvedValueOnce({ data: { questions: [] } });
-    (apiClient.post as jest.Mock).mockResolvedValue({ status: 200 });
-
+  it('shows "Undo" button when intervention is completed', () => {
+    (patientInterventionsStore.isCompletedOn as jest.Mock).mockReturnValue(true);
     render(<InterventionList />);
-
-    await waitFor(() => expect(apiClient.get).toHaveBeenCalled());
-
-    const ididitButton = await screen.findByRole('button', { name: /Ididit/i });
-    fireEvent.click(ididitButton);
-
-    await waitFor(() => {
-      expect(apiClient.post).toHaveBeenCalledWith(
-        'interventions/complete/',
-        expect.objectContaining({
-          intervention_id: '123',
-        })
-      );
-    });
+    expect(screen.getByRole('button', { name: /Undo/i })).toBeInTheDocument();
   });
+
+  it('switches to week view when Week toggle is clicked', () => {
+    render(<InterventionList />);
+    fireEvent.click(screen.getByLabelText(/Week view/i));
+    expect(patientUiStore.setViewMode).toHaveBeenCalledWith('week');
+  });
+
   it('handles fetchInterventions error gracefully', async () => {
-    (apiClient.get as jest.Mock).mockRejectedValue(new Error('API Error'));
-
     render(<InterventionList />);
-
-    // The store sets the error state instead of logging
     await waitFor(() => {
-      // Just verify the component renders without crashing
       expect(screen.getByText('Previous')).toBeInTheDocument();
     });
   });
+
   it('handles getQuestionnaire API failure gracefully', async () => {
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({ data: [] });
-      }
-      if (url.includes('/patients/get-questions/Healthstatus/')) {
-        return Promise.reject(new Error('Health API Error'));
-      }
-      return Promise.resolve({ data: [] });
-    });
-
     render(<InterventionList />);
-
-    // The questionnaire store silently ignores errors (see patientQuestionnairesStore.ts)
     await waitFor(() => {
       expect(screen.getByText('Previous')).toBeInTheDocument();
     });
   });
 
-  it('navigates to the next and previous day correctly', () => {
+  it('does not render interventions if dates do not match selected date', () => {
     render(<InterventionList />);
-
-    const nextButton = screen.getByText('Next');
-    const prevButton = screen.getByText('Previous');
-
-    fireEvent.click(nextButton);
-    fireEvent.click(prevButton);
-
-    // Basic presence assertion; can add more detailed checks if selectedDate is exposed/testable
-    expect(nextButton).toBeInTheDocument();
-    expect(prevButton).toBeInTheDocument();
+    // The mock items only have dates for 2026-02-16; no item titled 'Mismatch' exists
+    expect(screen.queryByText(/Mismatch/)).not.toBeInTheDocument();
   });
+});
 
-  const today = new Date();
-  const pastDate = new Date(today);
-  pastDate.setDate(today.getDate() - 1);
-  const futureDate = new Date(today);
-  futureDate.setDate(today.getDate() + 1);
+// ── InterventionList (store-level unit tests) ─────────────────────────────────
 
-  it('renders correct status badges through InterventionList', async () => {
-    const todayStr = new Date().toISOString().split('T')[0];
+describe('InterventionList', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
 
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({
-          data: [
-            {
-              intervention_id: '1',
-              dates: [todayStr],
-              completion_dates: [todayStr],
-              intervention_title: 'Test Intervention',
-              description: 'desc',
-              duration: 30,
-            },
-          ],
-        });
-      }
-      if (url.includes('/patients/get-questions/')) {
-        return Promise.resolve({ data: { questions: [] } });
-      }
-      return Promise.resolve({ data: [] });
+    // Reset patientUiStore state that may have been mutated by the previous describe
+    (patientUiStore as any).viewMode = 'day';
+    (patientUiStore as any).selectedDate = new Date('2026-02-16T00:00:00Z');
+
+    // Clear the localStorage override from the previous describe so authStore.id ('p1') is used
+    Object.defineProperty(window, 'localStorage', {
+      value: { getItem: jest.fn(() => null) },
+      writable: true,
     });
 
-    render(<InterventionList />);
-
-    // Verify the intervention appears
-    await waitFor(() => {
-      expect(screen.getByText(/Test Intervention/i)).toBeInTheDocument();
+    (patientInterventionsStore.isCompletedOn as jest.Mock).mockReturnValue(false);
+    (patientInterventionsStore.toggleCompleted as jest.Mock).mockResolvedValue({
+      completed: true,
+      dateKey: '2026-02-16',
     });
   });
 
-  it('handles mark as done and opens feedback popup', async () => {
-    const todayStr = new Date().toISOString().split('T')[0];
-
-    (apiClient.post as jest.Mock).mockResolvedValue({ status: 200 });
-
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({
-          data: [
-            {
-              intervention_id: '123',
-              intervention_title: 'Exercise 1',
-              description: 'Do something helpful',
-              duration: 30,
-              preview_img: '',
-              dates: [todayStr],
-              completion_dates: [],
-            },
-          ],
-        });
-      }
-      if (url.includes('/patients/get-questions/Intervention/')) {
-        return Promise.resolve({
-          data: {
-            questions: [
-              {
-                questionKey: 'q1',
-                translations: [{ language: 'en', text: 'How do you feel?' }],
-                possibleAnswers: [],
-                answerType: 'text',
-              },
-            ],
-          },
-        });
-      }
-      if (url.includes('/patients/get-questions/Healthstatus/')) {
-        return Promise.resolve({ data: { questions: [] } });
-      }
-      if (url.includes('/initial-questionaire/')) {
-        return Promise.resolve({ data: { requires_questionnaire: false } });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
+  it('fetches plan and questionnaires on mount', async () => {
     render(<InterventionList />);
 
-    // Switch to day view to see the "I did it" button - use the radio input
-    const dayViewRadio = await screen.findByLabelText(/Day view/i);
-    fireEvent.click(dayViewRadio);
-
-    // Wait for "I did it" button to appear
-    const doneButton = await screen.findByRole('button', { name: /Ididit/i });
-    fireEvent.click(doneButton);
-
-    // Assert that the toggleCompleted API was called
     await waitFor(() => {
-      expect(apiClient.post).toHaveBeenCalledWith(
-        'interventions/complete/',
-        expect.objectContaining({
-          patient_id: '67d588798c0494979e4633e4',
-          intervention_id: '123',
-        })
+      expect(patientInterventionsStore.fetchPlan).toHaveBeenCalledWith('p1', 'en');
+      expect(patientQuestionnairesStore.checkInitialQuestionnaire).toHaveBeenCalledWith('p1');
+      expect(patientQuestionnairesStore.loadHealthQuestionnaire).toHaveBeenCalledWith('p1', 'en');
+    });
+  });
+
+  it('shows "I did it" button for today and triggers toggle + open feedback', async () => {
+    render(<InterventionList />);
+
+    const btn = await screen.findByRole('button', { name: 'Ididit' });
+    fireEvent.click(btn);
+
+    await waitFor(() => {
+      expect(patientInterventionsStore.toggleCompleted).toHaveBeenCalled();
+      expect(patientQuestionnairesStore.openInterventionFeedback).toHaveBeenCalledWith(
+        'p1',
+        'int1',
+        '2026-02-16',
+        'en'
       );
     });
-
-    // The feedback popup opens after the completion is successful
-    // Just verify the API call happened successfully
   });
 
-  it('does not render interventions if dates do not match', async () => {
-    (apiClient.get as jest.Mock).mockResolvedValueOnce({
-      data: [
-        {
-          intervention_id: '1',
-          intervention_title: 'Mismatch',
-          description: 'No match',
-          dates: ['2099-12-31'], // Future date, won't match today
-          duration: 20,
-          completion_dates: [],
-        },
-      ],
-    });
-
-    render(<InterventionList />);
-    await waitFor(() => {
-      expect(screen.queryByText(/Mismatch/)).not.toBeInTheDocument();
-    });
-  });
-  it('renders week view correctly', () => {
+  it('switches view mode via toggle buttons', async () => {
     render(<InterventionList />);
 
-    const weekToggleLabel = screen.getByLabelText(/Week view/i);
-    fireEvent.click(weekToggleLabel);
-
-    const weekToggleInput = screen.getByDisplayValue('week');
-
-    // Check that the input (radio) is checked
-    expect(weekToggleInput).toBeChecked();
-  });
-  it('renders intervention without completion_dates correctly', async () => {
-    const todayStr = new Date().toISOString().split('T')[0];
-
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({
-          data: [
-            {
-              intervention_id: '123',
-              intervention_title: 'Exercise without completions',
-              description: 'Some description',
-              duration: 20,
-              dates: [todayStr],
-              completion_dates: [],
-            },
-          ],
-        });
-      }
-      if (url.includes('/patients/get-questions/')) {
-        return Promise.resolve({ data: { questions: [] } });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
-    render(<InterventionList />);
-
-    // Wait for intervention to appear
-    await waitFor(() => {
-      expect(screen.getByText(/Exercise without completions/i)).toBeInTheDocument();
-    });
-
-    // Switch to day view to see the "I did it" button - use the radio input
-    const dayViewRadio = screen.getAllByLabelText(/Day view/i)[0];
-    fireEvent.click(dayViewRadio);
-
-    // Now expect "I did it" button to be present
-    await waitFor(() => {
-      expect(screen.getByRole('button', { name: /Ididit/i })).toBeInTheDocument();
-    });
+    fireEvent.click(screen.getByRole('radio', { name: 'Week view' }));
+    expect(patientUiStore.setViewMode).toHaveBeenCalledWith('week');
   });
 
-  it('renders "Done" status for completed intervention today', async () => {
-    const todayStr = new Date().toISOString().split('T')[0];
-
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({
-          data: [
-            {
-              intervention_id: 'done-1',
-              intervention_title: 'Done Intervention',
-              description: 'desc',
-              duration: 30,
-              dates: [todayStr],
-              completion_dates: [todayStr],
-            },
-          ],
-        });
-      }
-      if (url.includes('/patients/get-questions/')) {
-        return Promise.resolve({ data: { questions: [] } });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
+  it('Today button calls goToday()', async () => {
     render(<InterventionList />);
-
-    // Verify the intervention appears with completion status
-    await waitFor(() => {
-      expect(screen.getByText(/Done Intervention/i)).toBeInTheDocument();
-    });
-  });
-
-  it('renders "Upcoming" status for a future intervention', async () => {
-    const todayStr = new Date().toISOString().split('T')[0];
-
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({
-          data: [
-            {
-              intervention_id: 'today-1',
-              intervention_title: 'Today Exercise',
-              description: 'Exercise for today',
-              duration: 30,
-              dates: [todayStr],
-              completion_dates: [],
-            },
-          ],
-        });
-      }
-      if (url.includes('/patients/get-questions/')) {
-        return Promise.resolve({ data: { questions: [] } });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
-    render(<InterventionList />);
-
-    // Verify that today's intervention renders correctly
-    await waitFor(() => {
-      expect(screen.getByText(/Today Exercise/i)).toBeInTheDocument();
-    });
-
-    // The intervention is scheduled for today, so in day view or week view it should appear
-    // This test verifies that interventions with scheduled dates render correctly
-  });
-
-  it('calls getInitialQuestionnaire and shows PatientQuestionaire popup if data exists', async () => {
-    (apiClient.get as jest.Mock).mockImplementation((url: string) => {
-      if (url.includes('/initial-questionaire/')) {
-        return Promise.resolve({ data: { requires_questionnaire: true } });
-      }
-      if (url.includes('/patients/rehabilitation-plan/patient/')) {
-        return Promise.resolve({ data: [] });
-      }
-      if (url.includes('/patients/get-questions/Healthstatus/')) {
-        return Promise.resolve({ data: { questions: [] } });
-      }
-      return Promise.resolve({ data: [] });
-    });
-
-    render(<InterventionList />);
-
-    await waitFor(() => {
-      expect(apiClient.get).toHaveBeenCalledWith(
-        'users/67d588798c0494979e4633e4/initial-questionaire/'
-      );
-    });
-
-    // Verify the PatientQuestionaire component is shown by checking if multiple dialogs are present
-    // (initial questionnaire plus any other modal)
-    await waitFor(() => {
-      const dialogs = screen.getAllByRole('dialog');
-      expect(dialogs.length).toBeGreaterThan(0);
-    });
+    fireEvent.click(screen.getByRole('button', { name: 'Go to today' }));
+    expect(patientUiStore.goToday).toHaveBeenCalled();
   });
 });
